pyproject.toml
[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"

[project]
name = "icenet-mp"
version = "0.2.0"
description = "IceNet Multimodal Pipeline for predicting sea ice."
readme = "README.md"
authors = [
    { name = "Sophie Arana", email = "arana.sophie@gmail.com" },
    { name = "Simon Driscoll", email = "simondriscoll88@gmail.com" },
    { name = "Isabel Fenton", email = "ifenton@turing.ac.uk" },
    { name = "Lydia France", email = "85945427+LydiaFrance@users.noreply.github.com" },
    { name = "Maria Carolina Novitasari", email ="124182131+marianovitasari20@users.noreply.github.com" },
    { name = "Nilo Pedrazzini", email = "npedrazzini@turing.ac.uk" },
    { name = "Erin Quan", email = "72717299+erinuclkwon@users.noreply.github.com" },
    { name = "James Robinson", email = "james.em.robinson@gmail.com" },
    { name = "Louisa van Zeeland", email = "louisa-ai2@protonmail.com" },
]
requires-python = ">=3.11,<3.13" # Note "3.13" is currently incompatible with the tensorflow requirement from icenet
dependencies = [
    "anemoi-datasets>=0.5.27", # to get the fix from https://github.com/ecmwf/anemoi-datasets/pull/372
    "anemoi-transform>=0.1.16",
    "anemoi-utils>=0.4.42",
    "argopy>=1.4.0",
    "cachetools>=6.1.0",
    "cdsapi>=0.7.6",
    "eccodeslib>=2.40.0,<=2.45.3,!=2.42.0,!=2.43.0,!=2.44.0.2", # indirect dependency of icenet that is not automatically installed
    "haversine>=2.9.0",
    "hydra-core>=1.3.2",
    "icenet==0.2.7", # newer versions require netCDF4<1.6.1 which does not work on macOS 15
    "imageio-ffmpeg>=0.6.0",
    "jaxtyping>=0.3.2",
    "lightning>=2.5.1",
    "matplotlib>=3.10.3",
    "netcdf4>=1.7.0",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pyproj>=3.7.2",
    "pytorch-triton-xpu>=0.0; sys_platform == 'linux' and platform_version == '#1 SMP PREEMPT_DYNAMIC Wed Dec 17 11:57:42 GMT 2025'", # only needed for XPU-enabled machines, but we don't have a simple way to determine that
    "pytorch-triton>=0.0; sys_platform == 'linux' and platform_version != '#1 SMP PREEMPT_DYNAMIC Wed Dec 17 11:57:42 GMT 2025'",
    "torch-ema>=0.3",
    "torch>=2.8.0,!=2.10.0", # uv will constantly reinstall torch 2.10.0 due to incorrect metadata https://github.com/astral-sh/uv/issues/17711
    "typer>=0.16.0",
    "wandb>=0.21.0",
    "xarray>=2025.3.0",
    "zarr<3.0.0", # v3 is not compatible with anemoi
]

[dependency-groups]
dev = [
    "mypy>=1.17.0",
    "pandas-stubs>=2.3.0.250703",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "types-cachetools>=6.1.0",
]
notebooks = [
    "cmocean>=4.0.3",
    "download-toolbox>=0.0.6",
    "jupyter>=1.1.1",
]

[project.scripts]
imp = "icenet_mp.cli:main"

[tool.coverage.paths]
source = ["icenet_mp/"]

[tool.coverage.run]
relative_files = true
omit= [
  "tests/*",
]

[[tool.mypy.overrides]]
module = [
    "anemoi.datasets.*",
    "anemoi.transform.*",
    "anemoi.utils.*",
    "earthkit.*",
    "icenet.data.sic.*",
    "imageio_ffmpeg.*",
    "zarr.*",
    "argopy.*",
    "haversine.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = [
    "--cov=icenet_mp",
    "--cov-report=term-missing",
    "-vvv",
]
testpaths = ["tests"]
norecursedirs = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
]

[tool.ruff]
line-length = 88
exclude = ["notebooks/*ipynb"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812", # conflicts with formatter
    "D100", # undocumented-public-module
    "D101", # undocumented-public-class
    "D102", # undocumented-public-method
    "D104", # undocumented-public-package
    "D203", # one-blank-line-before-class [conflicts with D211]
    "D213", # multi-line-summary-second-line [conflicts with D212]
    "D401", # non-imperative-mood
    "E501", # line-too-long
    "SLF001", # private-member-access
]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.isort]
known-third-party = ["wandb"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "DTZ001", # allow unlocalised datetimes in tests
    "S101", # allow asserts in tests
    "INP001", # allow implicit namespace in tests
    "PLR0913", # allow unlimited arguments in tests
    "PLR2004", # allow magic-number comparison in tests
]

[tool.ruff.lint.pylint]
max-args = 6

[tool.uv.build-backend]
module-root = ""

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-xpu"
url = "https://download.pytorch.org/whl/xpu"
explicit = true

[tool.uv.sources]
pytorch-triton = { index = "pytorch-cu128" }
pytorch-triton-xpu = { index = "pytorch-xpu" }
torch = [
    # IsambardAI
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' and platform_machine == 'aarch64'" },
    # Baskerville
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' and platform_machine == 'x86_64' and platform_version != '#1 SMP PREEMPT_DYNAMIC Wed Dec 17 11:57:42 GMT 2025'" },
    # DAWN
    { index = "pytorch-xpu", marker = "sys_platform == 'linux' and platform_machine == 'x86_64' and platform_version == '#1 SMP PREEMPT_DYNAMIC Wed Dec 17 11:57:42 GMT 2025'" },
]
triton = [
    { index = "pytorch-cu128", marker = "platform_version != '#1 SMP PREEMPT_DYNAMIC Wed Dec 17 11:57:42 GMT 2025'" },
    { index = "pytorch-xpu", marker = "platform_version == '#1 SMP PREEMPT_DYNAMIC Wed Dec 17 11:57:42 GMT 2025'" },
]
